[WIP] fix build

+++ sdl2-sonic-drivers/src/hardware/opl/scummvm/Config.cpp
@@ -41,7 +41,7 @@ namespace hardware
                     return std::make_shared<woody::WoodyOPL>(mixer, type == OplType::OPL2 ? false : true);
                     
                 default:
-                    spdlog::error("Unsupported OPL emulator {0:d}", oplEmulator);
+                    spdlog::error("Unsupported OPL emulator {0:d}", (int)oplEmulator);
                 }
 
                 return nullptr;
